requirements_new.txt
absl-py==2.1.0
accelerate==1.3.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.15.2
asttokens==3.0.0
attrs==24.3.0
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
coverage==7.6.10
DataProperty==1.1.0
datasets==2.20.0
debugpy==1.8.12
decorator==5.1.1
dill==0.3.8
docker-pycreds==0.4.0
evaluate==0.4.3
executing==2.2.0
filelock==3.17.0
frozenlist==1.5.0
fsspec
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.1
huggingface-hub==0.29.3
idna==3.10
iniconfig==2.0.0
inquirerpy==0.3.4
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonlines==4.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
lm_eval==0.4.3
lxml==5.3.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mbstrdecoder==1.1.4
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numexpr==2.10.2
numpy==1.26.4
optuna==4.3.0
packaging==24.2
pandas==2.2.2
parso==0.8.4
pathvalidate==3.2.3
peft==0.14.0
pexpect==4.9.0
pfzy==0.3.4
platformdirs==4.3.6
pluggy==1.5.0
portalocker==3.1.1
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.0
pyarrow-hotfix==0.6
pybind11==2.13.6
Pygments==2.19.1
pytablewriter==1.2.1
pytest==8.3.2
pytest-cov==6.0.0
pytest-subtests==0.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
rouge_score==0.1.2
sacrebleu==2.5.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.1
sentencepiece==0.2.0
sentry-sdk==2.20.0
setproctitle==1.3.4
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.40
sqlitedict==2.1.0
stack-data==0.6.3
steering-vectors==0.12.2
sympy==1.13.1
tabledata==1.3.4
tabulate==0.9.0
tcolorpy==0.1.7
threadpoolctl==3.5.0
tokenizers==0.21.1
torch
torchvision
torch-kmeans==0.2.0
tornado==6.4.2
tqdm==4.67.1
tqdm-multiprocess==0.0.11
traitlets==5.14.3
transformers
# triton==3.1.0
typepy==1.3.4
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wandb
wcwidth==0.2.13
word2number==1.1
xgboost==3.0.0
xxhash==3.5.0
yarl==1.18.3
zstandard==0.23.0